raft: moves id_allocator_stm from raft to cluster
