+++ src/v/cluster/CMakeLists.txt
@@ -48,6 +48,7 @@ v_cc_library(
     partition.cc
     partition_probe.cc
     seq_stm.cc
+    id_allocator_stm.cc
   DEPS
     Seastar::seastar
     controller_rpc
