+++ src/v/cluster/partition.h
@@ -11,6 +11,7 @@
 
 #pragma once
 
+#include "cluster/id_allocator_stm.h"
 #include "cluster/partition_probe.h"
 #include "cluster/seq_stm.h"
 #include "cluster/types.h"
@@ -18,7 +19,6 @@
 #include "model/record_batch_reader.h"
 #include "raft/consensus.h"
 #include "raft/group_configuration.h"
-#include "raft/id_allocator_stm.h"
 #include "raft/log_eviction_stm.h"
 #include "raft/types.h"
 #include "storage/types.h"
@@ -127,7 +127,7 @@ public:
         return _raft->get_latest_configuration_offset();
     }
 
-    ss::lw_shared_ptr<raft::id_allocator_stm>& id_allocator_stm() {
+    ss::lw_shared_ptr<cluster::id_allocator_stm>& id_allocator_stm() {
         return _id_allocator_stm;
     }
 
@@ -139,7 +139,7 @@ private:
 private:
     consensus_ptr _raft;
     ss::lw_shared_ptr<raft::log_eviction_stm> _nop_stm;
-    ss::lw_shared_ptr<raft::id_allocator_stm> _id_allocator_stm;
+    ss::lw_shared_ptr<cluster::id_allocator_stm> _id_allocator_stm;
     ss::shared_ptr<seq_stm> _seq_stm;
     ss::abort_source _as;
     partition_probe _probe;
